Create config file for the template engine
